Resolve #173
Actually there is some bug in an upstream, that breaks all chunked GET
requests. I've switched to POST method in all requests, even if it is
actually a get.

But there was still a funny bug, due to which all requests were GET.
Also, fixed an issue with checking connection to bap server.