introducing new python interface
The old one is still alive, but is moved to extras. The new interface
uses just popen, and gives an access to the whole project information.

It is high level, so it doesn't provide a direct interface to the
disassembler. For this the old plumbing interface, based on RPC server,
should be used. Due to the fact, that the old interface has lots of
dependencies, it is now made optional.