fixes parsing section and region statements
Note: this is a backport from bap-1.3.1 package, as this fix is
already there, but wasn't commited to the master branch of the
upstream repository.

The fix enables the special handling for the Section and Region types,
which use hex number without the 0x prefix. Ideally, we shouldn't
generate such input, but since historically this happened, we need to
make our parser robust enough to be able to chew such representation
also.